Rename platform to platforms
